refactor(dashboards-v2): de-duplicate panel cell + column-unit helpers
- collapse the per-panel stringifyCell copies onto the shared coerceToString
  util (list + table column builders)
- extract the column-unit resolver to Panels/utils/getColumnUnit; the table
  renderer and the editor's column options now share one implementation,
  aligning the editor onto the renderer's "empty entry = no unit" semantics
- export a single TableRowData type from tableColumns (with the antd key)
  instead of two divergent local copies
